cope all changes under kialayout
