+++ packages/easy-email-extensions/src/KiaLayout/Form/InputWithMerge/index.tsx
@@ -1,0 +1,8 @@
+import { InputProps } from '@arco-design/web-react';
+import { enhancer } from '@extensions/components/Form';
+import { InputWithMerge, InputWithMergeProps } from './InputWithMerge';
+
+export const TextWithMergeField = enhancer<
+InputWithMergeProps,
+Required<InputProps>['onChange']
+>(InputWithMerge, (value: string) => value);
